sum부분 LONG 시ㅏ도

+++ src/main/java/com/moaguide/dto/NewDto/customDto/CattleTransactionDto.java
@@ -1,21 +1,16 @@
 package com.moaguide.dto.NewDto.customDto;
 
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
-import java.math.BigInteger;
-import java.sql.Date;
 import java.time.LocalDate;
 
 @Getter
 @NoArgsConstructor
 public class CattleTransactionDto {
     private LocalDate day;
-    private BigDecimal value;
+    private Long  value;
 
-    public CattleTransactionDto(LocalDate date, BigDecimal value) {
+    public CattleTransactionDto(LocalDate date, Long  value) {
         this.day = date;
         this.value = value;
     }
